Combine ManifestReaders into one parameterized class. (#1099)
