+++ core/src/main/java/org/apache/iceberg/RemoveSnapshots.java
@@ -22,7 +22,6 @@ package org.apache.iceberg;
 import java.io.IOException;
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Consumer;
@@ -31,7 +30,6 @@ import org.apache.iceberg.exceptions.CommitFailedException;
 import org.apache.iceberg.exceptions.NotFoundException;
 import org.apache.iceberg.exceptions.RuntimeIOException;
 import org.apache.iceberg.io.CloseableIterable;
-import org.apache.iceberg.io.FileIO;
 import org.apache.iceberg.relocated.com.google.common.base.Joiner;
 import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
@@ -351,7 +349,7 @@ class RemoveSnapshots implements ExpireSnapshots {
         .onFailure((item, exc) -> LOG.warn("Failed to get deleted files: this may cause orphaned data files", exc))
         .run(manifest -> {
           // the manifest has deletes, scan it to find files to delete
-          try (BaseManifestReader<?, ?> reader = openManifest(manifest, ops.io(), ops.current().specsById())) {
+          try (ManifestReader<?> reader = ManifestFiles.open(manifest, ops.io(), ops.current().specsById())) {
             for (ManifestEntry<?> entry : reader.entries()) {
               // if the snapshot ID of the DELETE entry is no longer valid, the data can be deleted
               if (entry.status() == ManifestEntry.Status.DELETED &&
@@ -371,7 +369,7 @@ class RemoveSnapshots implements ExpireSnapshots {
         .onFailure((item, exc) -> LOG.warn("Failed to get added files: this may cause orphaned data files", exc))
         .run(manifest -> {
           // the manifest has deletes, scan it to find files to delete
-          try (ManifestReader reader = ManifestFiles.read(manifest, ops.io(), ops.current().specsById())) {
+          try (ManifestReader<?> reader = ManifestFiles.open(manifest, ops.io(), ops.current().specsById())) {
             for (ManifestEntry<?> entry : reader.entries()) {
               // delete any ADDED file from manifests that were reverted
               if (entry.status() == ManifestEntry.Status.ADDED) {
@@ -402,16 +400,5 @@ class RemoveSnapshots implements ExpireSnapshots {
     } else {
       return CloseableIterable.withNoopClose(snapshot.allManifests());
     }
-  }
-
-  private static BaseManifestReader<?, ?> openManifest(ManifestFile manifest, FileIO io,
-                                                       Map<Integer, PartitionSpec> specsById) {
-    switch (manifest.content()) {
-      case DATA:
-        return ManifestFiles.read(manifest, io, specsById);
-      case DELETES:
-        return ManifestFiles.readDeleteManifest(manifest, io, specsById);
-    }
-    throw new UnsupportedOperationException("Cannot read unknown manifest type: " + manifest.content());
   }
 }
