+++ core/src/test/java/org/apache/iceberg/TestTransaction.java
@@ -502,7 +502,7 @@ public class TestTransaction extends TableTestBase {
 
     // create a manifest append
     OutputFile manifestLocation = Files.localOutput("/tmp/" + UUID.randomUUID().toString() + ".avro");
-    ManifestWriter writer = ManifestFiles.write(table.spec(), manifestLocation);
+    ManifestWriter<DataFile> writer = ManifestFiles.write(table.spec(), manifestLocation);
     try {
       writer.add(FILE_D);
     } finally {
